deploy/requirements.txt
torch>=2.0
numpy
pyyaml
dynamixel-sdk
pyserial
